requirements.txt
matplotlib==3.5.2
pandas==1.4.4
Pillow==9.2.0
Pillow==9.5.0
Requests==2.31.0
scikit_learn==1.0.2
